Tratamento de erros implementada

+++ src/main/java/br/ifma/edu/imobiliaria/domain/service/ImobiliariaService.java
@@ -6,6 +6,7 @@ import javax.transaction.Transactional;
 
 import org.springframework.stereotype.Service;
 
+import br.ifma.edu.imobiliaria.domain.exception.NegocioException;
 import br.ifma.edu.imobiliaria.domain.model.Imobiliaria;
 import br.ifma.edu.imobiliaria.domain.repository.ImobiliariaRepository;
 import lombok.AllArgsConstructor;
@@ -16,10 +17,14 @@ public class ImobiliariaService {
     private final ImobiliariaRepository repository;
 
     public Iterable<Imobiliaria> todos() {
+        if (repository.findAll() == null)
+            throw new NegocioException("Não existe nenhum imovel cadastrado.");
         return repository.findAll();
     }
 
     public Optional<Imobiliaria> buscaPor(Long id) {
+        if (repository.findById(id) == null)
+            throw new NegocioException("Não existe nenhuma imobiliaria cadastrado com este id.");
         return repository.findById(id);
     }
 
@@ -30,6 +35,8 @@ public class ImobiliariaService {
 
     @Transactional
     public void removePelo(Long id) {
+        if (repository.findById(id) == null)
+            throw new NegocioException("Não existe nenhuma imobiliaria cadastrado com este id.");
         repository.deleteById(id);
     }
